+++ src/js/handlers/authorise-google.js
@@ -1,5 +1,5 @@
-var fireGAEvent = require( 'helpers/fireGAEvent' );
-var storeLastUsedConnection = require( 'helpers/storeLastUsedConnection' );
+var fireGAEvent = require( 'helpers/fire-ga-event' );
+var storeLastUsedConnection = require( 'helpers/store-last-used-connection' );
 
 module.exports = function authorise( element ) {
   var form = element.closest( 'form' );
