+++ src/js/handlers/authorise-ldap.js
@@ -1,6 +1,6 @@
 var ui = require( 'helpers/ui' );
-var fireGAEvent = require( 'helpers/fireGAEvent' );
-var storeLastUsedConnection = require( 'helpers/storeLastUsedConnection' );
+var fireGAEvent = require( 'helpers/fire-ga-event' );
+var storeLastUsedConnection = require( 'helpers/store-last-used-connection' );
 
 module.exports = function authorise( element, secondTry ) {
   var form = element.tagName === 'FORM' ? element : element.form;
