+++ src/js/handlers/go-to-initial-page.js
@@ -1,5 +1,5 @@
 var ui = require( 'helpers/ui' );
-var fireGAEvent = require( 'helpers/fireGAEvent' );
+var fireGAEvent = require( 'helpers/fire-ga-event' );
 
 module.exports = function goToInitialPage( element ) {
   ui.setLockState( element, 'initial' );
