+++ src/js/helpers/store-last-used-connection.js
